+++ src/mesh3/examples/integration/node_test.go
@@ -1,4 +1,4 @@
-package main
+package integration2
 
 import (
 	"bytes"
@@ -12,9 +12,8 @@ import (
 
 	"github.com/satori/go.uuid"
 	"github.com/skycoin/skycoin/src/cipher"
-	mesh "github.com/skycoin/skycoin/src/mesh3/node"
-	"github.com/skycoin/skycoin/src/mesh3/transport/reliable"
-	"github.com/skycoin/skycoin/src/mesh3/transport/udp"
+	"github.com/skycoin/skycoin/src/mesh2"
+	"github.com/skycoin/skycoin/src/mesh3/transport/protocol"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -41,8 +40,8 @@ type ToConnect struct {
 }
 
 type TestConfig struct {
-	Reliable reliable.ReliableTransportConfig
-	Udp      udp.UDPConfig
+	Reliable protocol.ReliableTransportConfig
+	Udp      protocol.UDPConfig
 	Node     mesh.NodeConfig
 
 	PeersToConnect    []ToConnect
@@ -158,8 +157,8 @@ func createTestConfig(configText string) TestConfig {
 }
 
 // Create UDPTransport
-func createNewUDPTransport(configUdp udp.UDPConfig) *udp.UDPTransport {
-	udpTransport, createUDPError := udp.NewUDPTransport(configUdp)
+func createNewUDPTransport(configUdp protocol.UDPConfig) *protocol.UDPTransport {
+	udpTransport, createUDPError := protocol.NewUDPTransport(configUdp)
 	if createUDPError != nil {
 		panic(createUDPError)
 	}
@@ -170,8 +169,8 @@ func createNewUDPTransport(configUdp udp.UDPConfig) *udp.UDPTransport {
 func createTestConfig2(port int) TestConfig {
 	testConfig := TestConfig{}
 	testConfig.Node = mesh.NewNodeConfig()
-	testConfig.Reliable = reliable.CreateReliable(testConfig.Node.PubKey)
-	testConfig.Udp = udp.CreateUdp(port, "127.0.0.1")
+	testConfig.Reliable = protocol.CreateReliable(testConfig.Node.PubKey)
+	testConfig.Udp = protocol.CreateUdp(port, "127.0.0.1")
 
 	return testConfig
 }
@@ -186,8 +185,7 @@ func TestSendMessage(t *testing.T) {
 	peersToConnect1 := []ToConnect{}
 	peerToConnect1 := ToConnect{}
 	peerToConnect1.Peer = config2.Node.PubKey
-	cryptoKey1 := []byte{1, 0, 0, 0, 1, 0, 44, 22, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 11, 0, 0}
-	peerToConnect1.Info = udp.CreateUDPCommConfig("127.0.0.1:17000", cryptoKey1[:])
+	peerToConnect1.Info = protocol.CreateUDPCommConfig("127.0.0.1:17000", config2.Node.ChaCha20Key[:])
 	peersToConnect1 = append(peersToConnect1, peerToConnect1)
 	config1.PeersToConnect = peersToConnect1
 
@@ -215,8 +213,7 @@ func TestSendMessage(t *testing.T) {
 	peersToConnect2 := []ToConnect{}
 	peerToConnect2 := ToConnect{}
 	peerToConnect2.Peer = config1.Node.PubKey
-	cryptoKey2 := []byte{1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 11, 22, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0}
-	peerToConnect2.Info = udp.CreateUDPCommConfig("127.0.0.1:15000", cryptoKey2[:])
+	peerToConnect2.Info = protocol.CreateUDPCommConfig("127.0.0.1:15000", config1.Node.ChaCha20Key[:])
 	peersToConnect2 = append(peersToConnect2, peerToConnect2)
 	config2.PeersToConnect = peersToConnect2
 
@@ -277,7 +274,7 @@ func TestPubKey(t *testing.T) {
 // Validates that info to peer connect is equal.
 func TestUDPCommConfig(t *testing.T) {
 	cryptoKey := []byte{1, 55, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0}
-	enc := udp.CreateUDPCommConfig("127.0.0.1:16000", cryptoKey)
+	enc := protocol.CreateUDPCommConfig("127.0.0.1:16000", cryptoKey)
 
 	expected := "7b22446174616772616d4c656e677468223a3531322c2245787465726e616c486f737473223a5b7b224950223a223132372e302e302e31222c22506f7274223a31363030302c225a6f6e65223a22227d5d2c2243727970746f4b6579223a22415463414141454141414142414141414151414141414541414141424141414141514141414145414141413d227d"
 	assert.Equal(t, expected, enc, "Error in encoding")
@@ -298,13 +295,13 @@ func TestNodeCase1(t *testing.T) {
 	// Initialize Node 2
 	config2 := createTestConfig(configText2)
 	cryptoKey2 := []byte{1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 11, 22, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0}
-	config2.PeersToConnect[0].Info = udp.CreateUDPCommConfig("127.0.0.1:15000", cryptoKey2)
+	config2.PeersToConnect[0].Info = protocol.CreateUDPCommConfig("127.0.0.1:15000", cryptoKey2)
 	go InitializeNode(2, config2, &wg, statusChannel)
 
 	// Initialize Node 1
 	config1 := createTestConfig(configText1)
 	cryptoKey1 := []byte{1, 0, 0, 0, 1, 0, 44, 22, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 11, 0, 0}
-	config1.PeersToConnect[0].Info = udp.CreateUDPCommConfig("127.0.0.1:17000", cryptoKey1)
+	config1.PeersToConnect[0].Info = protocol.CreateUDPCommConfig("127.0.0.1:17000", cryptoKey1)
 	go InitializeNode(1, config1, &wg, statusChannel)
 
 	timeout := 30 * time.Second
@@ -343,7 +340,7 @@ func InitializeNode(idConfig int, config TestConfig, wg *sync.WaitGroup, statusC
 	}
 
 	// Reliable transport closes UDPTransport
-	reliableTransport := reliable.NewReliableTransport(udpTransport, config.Reliable)
+	reliableTransport := protocol.NewReliableTransport(udpTransport, config.Reliable)
 	defer reliableTransport.Close()
 
 	node, createNodeError := mesh.NewNode(config.Node)
