+++ src/mesh3/node/node.go
