+++ src/mesh3/node/node_test.go
