+++ src/mesh3/serialize/serialize.go
