+++ src/mesh3/serialize/serialize_test.go
