+++ src/mesh3/transport/protocol/reliable.go
@@ -1,4 +1,4 @@
-package reliable
+package protocol
 
 import (
 	"fmt"
@@ -8,8 +8,8 @@ import (
 	"time"
 
 	"github.com/skycoin/skycoin/src/cipher"
-	"github.com/skycoin/skycoin/src/mesh3/serialize"
-	"github.com/skycoin/skycoin/src/mesh3/transport/transport"
+	"github.com/skycoin/skycoin/src/mesh2/serialize"
+	"github.com/skycoin/skycoin/src/mesh2/transport/transport"
 )
 
 import (
