+++ src/mesh3/transport/protocol/reliable_test.go
@@ -1,4 +1,4 @@
-package reliable
+package protocol
 
 import (
 	"testing"
@@ -7,7 +7,7 @@ import (
 
 import (
 	"github.com/skycoin/skycoin/src/cipher"
-	"github.com/skycoin/skycoin/src/mesh3/transport/transport"
+	"github.com/skycoin/skycoin/src/mesh2/transport/transport"
 	"github.com/stretchr/testify/assert"
 )
