+++ src/mesh3/transport/protocol/udp.go
@@ -1,4 +1,4 @@
-package udp
+package protocol
 
 import (
 	"crypto/rand"
@@ -16,7 +16,7 @@ import (
 
 	"github.com/skycoin/skycoin/src/cipher"
 	"github.com/skycoin/skycoin/src/cipher/encoder"
-	"github.com/skycoin/skycoin/src/mesh3/transport/transport"
+	"github.com/skycoin/skycoin/src/mesh2/transport/transport"
 )
 
 import (
